+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enforce convenience features such as documentation for public functionality and the existence of debug formats for all public data structures.
 - Start a changelog documenting all relevant changes throughout versions.
 - Mechanism for managing keys representing variables (`rusttyc::TcVar`) inside RustTyC rather than requiring the user to do so.
+- Variadic keys and AbstractTypes
 
 ### Fixed
 - Some outdated documentation.
@@ -16,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Leftover `dgb!`s.
 - Ena types from public API.
+- `get_type` removed because it cannot be used during the type check procedure as intended.
 
 ### Rename
 - `TypeChecker::new_key(&mut self) -> TypeCheckKey` has been renamed to `TypeChecker::new_term_key(&mut self) -> TypeCheckKey` to distinguish it from the newly added `TypeChecker::new_var_key(&mut self, var: Var) -> TypeCheckKey`.